Expose typed backend to `airfoil`
