+++ src/sketch/mod.rs
@@ -1,8 +1,11 @@
+/// Typedefs
+pub mod sketch;
+
 pub mod extrudes;
 pub mod hershey;
 pub mod image;
 pub mod metaballs;
 pub mod offset;
 pub mod shapes;
-pub mod sketch;
 pub mod truetype;
+pub mod airfoil;
